Increased height of notification dashboard
This commit increases the height of the notification table view
for better notification visibility as the button text to return
to the homescreen is no longer in place.

Even though it makes the button target smaller for the user, it's
still not difficult to tap the surrounding black area to return to
the homescreen.