Reworked dashboard animation + element positions
This commit reworks the animation that occurs when hiding / showing
the notification dashboard from a double press of the home button.

It also positions the UIActionSheet for clearing out all
notifications at the very bottom of the screen, overlaying the
multitasking tray.

Since the dialog is modal and destructive in nature, it should be
the only user manipulatable object on screen at that point in time.