Sped up and smoothed show/hide animations
This commit applies the animation of opacity and position to
only the containing UIWindow, not each individual element contained
within it. This dramatically smoothens the animations as it reduces
the work required by the device.

It also sets the opacity of each inner element during creation
instead of during animation. The containing UIWindow is hidden
anyway, which allows for the opacity attributes to be set without
them becoming visible.