Simplified dashboard table and cell backgrounds
This commit rips out alertListViewBackground from the dashboard for
better performance and style.

This saves from having to allocate memory and perform an transform
during dashboard show/hide and improves overall rendering speed. As
a side note, the background pattern was so subtle it was hardly
noticeable anyways unless the screen is a full brightness.

This commit also improves how subviews are added to UITableViewCell's