Lockscreen and Emergency call screen fixes
When the pending notifications list is shown and the user attempts to
unlock with the security code keyboard, the list is immediately hidden
upon swiping to unlock.

When pending notifications list and/or notifications area are shown, and
the user moves to the emergency call screen, both are immediately
hidden.